# back-end/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Dhumi Demo Production API
  version: 0.1.0
  description: |
    Accepted mock-first implementation contract for the Dhumi Marketplace Dataset
    and Scraper Library demo. Customers call only Dhumi. Bright Data credentials,
    dataset IDs, snapshot IDs and raw errors are never public fields.
    Customer operations use browser sessions. The 0070 refactor removes the
    three /v1/keys operations and secret-response recovery.

    This contract has no activation, customer plan, subscription, billing,
    credit or quota endpoints. Organization, verification and member workflows follow 0071.
    Live Bright Data traffic remains controlled by external evidence gates.
  contact:
    name: Dhumi Engineering
  license:
    name: Proprietary - internal implementation contract
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.dhumi.example
    description: Reserved documentation placeholder; deployment base URL is pending.
...
paths:
  /v1/auth/signup:
    post:
      tags: [Authentication]
      operationId: signUp
      x-dhumi-access: public
      summary: Submit a local Dhumi signup
      description: |
        Creates only the local User, legal evidence and audit atomically.
        Creates no organization, membership, OTP or notification outbox.
        workspace_name is optional, deprecated and ignored. Makes no Bright Data request.
        Response is deliberately generic. An existing email follows the same
        accepted flow; 409 is reserved for the same Idempotency-Key used with a
        different canonical request body.
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpInput'
      responses:
        '202':
          description: Signup submission accepted
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthAccepted'
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: "Same Idempotency-Key was used with a different canonical request (IDEMPOTENCY_CONFLICT); never a duplicate-email signal"
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /v1/auth/sign-in:
    post:
      tags: [Authentication]
      operationId: signIn
      x-dhumi-access: public
      summary: Sign in to Dhumi
      description: Creates a user/session token family without requiring organization membership.
      security: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInInput'
      responses:
        '200':
          description: Authenticated
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /v1/auth/refresh:
    post:
      tags: [Authentication]
      operationId: refreshSession
      x-dhumi-access: refresh_cookie_csrf
      summary: Rotate the browser session token family
      description: Refresh token is supplied through the approved secure cookie. Organization membership is not required.
      security: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Session refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /v1/auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      x-dhumi-access: browser_session
      summary: Revoke the current browser session family
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '204':
          description: Signed out
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /v1/services:
    get:
      tags: [Services]
      operationId: listServices
      x-dhumi-access: organization_member
      summary: List tenant-owned services
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/OrganizationSelector'
      responses:
        '404': { $ref: '#/components/responses/NotFound' }
        '200':
          description: Paginated services
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicePage'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
    post:
      tags: [Services]
      operationId: createService
      x-dhumi-access: organization_member
      summary: Create a tenant-owned service configuration
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/OrganizationSelector'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCreateInput'
      responses:
        '201':
          description: Service created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/services/{service_id}:
    get:
      tags: [Services]
      operationId: getService
      x-dhumi-access: organization_member
      summary: Get one tenant-owned service
      security:
        - BrowserBearer: []
      parameters:
        - $ref: '#/components/parameters/ServiceId'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/OrganizationSelector'
      responses:
        '400': { $ref: '#/components/responses/BadRequest' }
        '200':
          description: Service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  securitySchemes:
    BrowserBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived Dhumi browser access token.
  headers:
    RequestId:
      description: Dhumi request correlation ID
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
    RetryAfter:
      description: Seconds before a safe retry may be attempted
      schema:
        type: integer
        minimum: 1
  parameters:
    OrganizationSelector:
      name: X-Dhumi-Organization
      in: header
      required: false
      description: Selector checked against active membership. Organization operations require selection with several memberships (400 ORGANIZATION_REQUIRED), and reject no membership (403 ORGANIZATION_MEMBERSHIP_REQUIRED). Browse needs no selector or membership for all-access templates. An inaccessible supplied selector gives 404 RESOURCE_NOT_FOUND.
      schema: { type: string, format: uuid }
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: Optional caller correlation ID; Dhumi returns an accepted/generated ID.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Opaque caller key scoped by authenticated actor/Tenant and operation.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Required for state-changing browser session operations.
      schema:
        type: string
        minLength: 16
        maxLength: 512
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned by the previous page.
      schema:
        type: string
        minLength: 1
        maxLength: 2048
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    ServiceId:
      name: service_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    SignUpInput:
      type: object
      additionalProperties: false
      required: [email, password, legal_acceptances]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 256
        workspace_name:
          type: string
          maxLength: 120
          deprecated: true
          description: Ignored compatibility field. Signup creates no organization.
        legal_acceptances:
          type: array
          minItems: 1
          maxItems: 10
          items:
            $ref: '#/components/schemas/LegalAcceptanceInput'
    LegalAcceptanceInput:
      type: object
      additionalProperties: false
      required: [document_type, document_version, content_hash, accepted]
      properties:
        document_type:
          type: string
          minLength: 1
          maxLength: 64
        document_version:
          type: string
          minLength: 1
          maxLength: 64
        content_hash:
          type: string
          pattern: '^[A-Fa-f0-9]{64}$'
        accepted:
          const: true
    AuthAccepted:
      type: object
      additionalProperties: false
      required: [accepted, message]
      properties:
        accepted: { type: boolean, const: true, readOnly: true }
        message:
          type: string
          readOnly: true
          example: Account request accepted. Sign in to continue.
    SignInInput:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, format: password, minLength: 1, maxLength: 256 }
    AuthSession:
      type: object
      additionalProperties: false
      required: [access_token, token_type, expires_in, csrf_token]
      properties:
        access_token: { type: string, readOnly: true }
        token_type: { type: string, const: Bearer, readOnly: true }
        expires_in: { type: integer, minimum: 1, readOnly: true }
        csrf_token: { type: string, readOnly: true }
...
    ProductFamily:
      type: string
      enum: [marketplace_dataset, scraper_library]
...
    ServiceCreateInput:
      type: object
      additionalProperties: false
      required: [template_slug, name, configuration]
      properties:
        template_slug:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        name: { type: string, minLength: 1, maxLength: 120 }
        configuration:
          type: object
          additionalProperties: true
          description: Validated against the selected Template configuration_schema. Per-Run input is supplied only when a Run is created.
    Service:
      type: object
      additionalProperties: false
      required: [id, name, template_slug, template_version, version, family, state, created_at]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        name: { type: string, readOnly: true }
        template_slug: { type: string, readOnly: true }
        template_version: { type: integer, minimum: 1, readOnly: true }
        version: { type: integer, minimum: 1, readOnly: true }
        family: { $ref: '#/components/schemas/ProductFamily' }
        state:
          type: string
          enum: [active, disabled]
          readOnly: true
        configuration:
          type: object
          readOnly: true
          additionalProperties: true
        created_at: { type: string, format: date-time, readOnly: true }
...
    Page:
      type: object
      additionalProperties: false
      required: [next_cursor, has_more]
      properties:
        next_cursor: { type: [string, 'null'], readOnly: true }
        has_more: { type: boolean, readOnly: true }
...
    ServicePage:
      type: object
      additionalProperties: false
      required: [data, page]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Service' } }
        page: { $ref: '#/components/schemas/Page' }
...
    ProblemField:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        field: { type: [string, 'null'] }
        message: { type: string }
    ProblemCode:
      type: string
      enum:
        - BAD_REQUEST
        - AUTHENTICATION_REQUIRED
        - ACCESS_DENIED
        - ORGANIZATION_REQUIRED
        - ORGANIZATION_MEMBERSHIP_REQUIRED
        - RESOURCE_NOT_FOUND
        - IDEMPOTENCY_CONFLICT
        - STATE_CONFLICT
        - PAYLOAD_TOO_LARGE
        - UNSUPPORTED_MEDIA_TYPE
        - VALIDATION_ERROR
        - SERVICE_INPUT_INVALID
        - PLATFORM_CAPACITY_LIMIT
        - INTERNAL_ERROR
        - SERVICE_UNAVAILABLE
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, request_id]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: [string, 'null'] }
        instance: { type: [string, 'null'], format: uri-reference }
        code: { $ref: '#/components/schemas/ProblemCode' }
        request_id: { type: string }
        errors:
          type: array
          items: { $ref: '#/components/schemas/ProblemField' }
...
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PayloadTooLarge:
      description: Request body exceeds the accepted one-mebibyte limit
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    UnsupportedMediaType:
      description: Request body does not use a supported media type
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid Dhumi authentication
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated but not authorized or Tenant unavailable
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource does not exist or is not visible to this Tenant
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Idempotency or current-state conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: Semantically invalid request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    TooManyRequests:
      description: Temporary Dhumi platform/product capacity limit; not a customer quota
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    InternalServerError:
      description: Unexpected Dhumi server error; internal detail is never exposed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ServiceUnavailable:
      description: Dhumi or enabled product dependency temporarily unavailable
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }